feat: implement OAF part 5 schemas: split test to own file
