+++ assets/i18n/en.yaml
@@ -108,7 +108,7 @@ ViewIn: View in the
 Browse: Browse through the
 BrowseSuffix: or go straight to the features in
 BrowseSchema: View the
-BrowseSchemaSuffix: of the features for mer details
+BrowseSchemaSuffix: of the features for more details
 FeaturesExplanation: GeoJSON is an open and widely accepted format for feature data and the first choice for many (web) applications. However, it formally only supports the WGS84 projection. JSON-FG is an extension of GeoJSON and does offer official support for other projections. Since JSON-FG is a recent development, from a compatibility point of view GeoJSON is also offered in other projections for the time being.
 CollectionTilesHTML: This collection is available as vector tiles in one or multiple projections.
 CollectionTilesPlain: This collection is available as vector tiles in one or multiple projections.
